Update to add toggles
Updated to make the labels and icons work in open and closed toggles. This file IS A BIG MESS. I am not a coder. Feel free to either use javascript or clean up my horribly inefficient php everywhere. I only wanted to get it working as a test, and it's crazy but it works.